Moved GradeTable component to its own module, imported into app
