+++ client/components/app.jsx
@@ -1,28 +1,7 @@
 import React from 'react';
 import Header from './header';
-import Grade from './grade';
+import GradeTable from './grade-table';
 import GradeForm from './grade-form';
-
-function GradeTable(props) {
-  const grades = props.grades;
-  const gradeRows = grades.map(grade => <Grade key={grade.id} grade={grade}/>);
-  return (
-    <div className="table-container mt-4 col-xl-8">
-      <table className="table table-striped">
-        <thead className="thead-light">
-          <tr>
-            <th scope="col">Student Name</th>
-            <th scope="col">Course</th>
-            <th scope="col">Grade</th>
-          </tr>
-        </thead>
-        <tbody>
-          { gradeRows }
-        </tbody>
-      </table>
-    </div>
-  );
-}
 
 class App extends React.Component {
   constructor(props) {
